initial separation of the HTL unit classes
