+++ exposan/htl/units_to_move/_struvite_precipitation.py
@@ -1,0 +1,151 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+'''
+QSDsan: Quantitative Sustainable Design for sanitation and resource recovery systems
+
+This module is developed by:
+
+    Jianan Feng <jiananf2@illinois.edu>
+
+    Yalin Li <mailto.yalin.li@gmail.com>
+
+This module is under the University of Illinois/NCSA Open Source License.
+Please refer to https://github.com/QSD-Group/QSDsan/blob/main/LICENSE.txt
+for license details.
+'''
+
+from math import ceil
+from qsdsan import SanUnit
+from . import Reactor
+
+__all__ = ('StruvitePrecipitation',)
+
+
+# =============================================================================
+# Struvite Precipitation
+# =============================================================================
+
+class StruvitePrecipitation(Reactor):
+    '''
+    Extracted and HTL aqueous are mixed together before adding MgCl2 for struvite precipitation.
+    If mol(N)<mol(P), add NH4Cl to mol(N):mol(P)=1:1.
+    
+    Parameters
+    ----------
+    ins : Iterable(stream)
+        mixture, supply_MgCl2, supply_NH4Cl, base.
+    outs : Iterable(stream)
+        struvite, effluent.
+    target_pH: float
+        Target pH for struvite precipitation.
+    Mg_P_ratio: float
+        mol(Mg) to mol(P) ratio.   
+    P_pre_recovery_ratio: float
+        Ratio of phosphorus that can be precipitated out.
+    HTLaqueous_NH3_N_2_total_N: float
+        Ratio of NH3-N to TN in HTL aqueous phase.
+        
+    References
+    ----------
+    .. [1] Zhu, Y.; Schmidt, A.; Valdez, P.; Snowden-Swan, L.; Edmundson, S.
+        Hydrothermal Liquefaction and Upgrading of Wastewater-Grown Microalgae:
+        2021 State of Technology; PNNL-32695, 1855835; 2022; p PNNL-32695, 1855835.
+        https://doi.org/10.2172/1855835.
+    .. [2] Jones, S. B.; Zhu, Y.; Anderson, D. B.; Hallen, R. T.; Elliott, D. C.; 
+        Schmidt, A. J.; Albrecht, K. O.; Hart, T. R.; Butcher, M. G.; Drennan, C.; 
+        Snowden-Swan, L. J.; Davis, R.; Kinchin, C. 
+        Process Design and Economics for the Conversion of Algal Biomass to
+        Hydrocarbons: Whole Algae Hydrothermal Liquefaction and Upgrading;
+        PNNL--23227, 1126336; 2014; https://doi.org/10.2172/1126336.
+    '''
+    _N_ins = 4
+    _N_outs = 2
+    _F_BM_default = {**Reactor._F_BM_default}
+    
+    def __init__(self, ID='', ins=None, outs=(), thermo=None,
+                 init_with='WasteStream', 
+                 target_pH = 9,
+                 Mg_P_ratio=1,
+                 P_pre_recovery_ratio=0.828, # [1]
+                 HTLaqueous_NH3_N_2_total_N = 0.853, # [2]
+                 P=None, tau=1, V_wf=0.8, # tau: [1]
+                 length_to_diameter=2, N=1, V=20, auxiliary=False,
+                 mixing_intensity=None, kW_per_m3=0, # use MixTank default value
+                 wall_thickness_factor=1,
+                 vessel_material='Carbon steel', # basic condition
+                 vessel_type='Vertical'):
+        
+        SanUnit.__init__(self, ID, ins, outs, thermo, init_with)
+        self.target_pH = target_pH
+        self.Mg_P_ratio = Mg_P_ratio
+        self.P_pre_recovery_ratio = P_pre_recovery_ratio
+        self.HTLaqueous_NH3_N_2_total_N = HTLaqueous_NH3_N_2_total_N
+        self.P = P
+        self.tau = tau
+        self.V_wf = V_wf
+        self.length_to_diameter = length_to_diameter
+        self.N = N
+        self.V = V
+        self.auxiliary = auxiliary
+        self.mixing_intensity = mixing_intensity
+        self.kW_per_m3 = kW_per_m3
+        self.wall_thickness_factor = wall_thickness_factor
+        self.vessel_material = vessel_material
+        self.vessel_type = vessel_type
+    
+        
+    def _run(self):
+        
+        mixture, supply_MgCl2, supply_NH4Cl, base = self.ins
+        struvite, effluent = self.outs
+        
+        self.HTLmixer = self.ins[0]._source
+        
+        if self.HTLmixer.outs[0].imass['P'] == 0:
+            effluent.copy_like(mixture)
+        else:
+            old_pH = self.HTLmixer.pH
+            neutral_OH_mol = 10**(-old_pH)*self.ins[0].F_mass # ignore solid volume
+            to_target_pH = 10**(self.target_pH - 14)*self.ins[0].F_mass # ignore solid volume
+            total_OH = neutral_OH_mol + to_target_pH # unit: mol/h
+            base.imass['MgO'] = total_OH/2 * 40.3044/1000
+            
+            supply_MgCl2.imass['MgCl2'] = max((mixture.imass['P']/30.973762*self.Mg_P_ratio -\
+                                           base.imass['MgO']/40.3044)*95.211, 0)
+    
+            if mixture.imass['P']/30.973762 > mixture.imass['N']*self.HTLaqueous_NH3_N_2_total_N/14.0067:
+            # if P > N, add NH4Cl to make sure N ≥ P
+                supply_NH4Cl.imass['NH4Cl'] = (mixture.imass['P']/30.973762 - mixture.imass['N']*\
+                                               self.HTLaqueous_NH3_N_2_total_N/14.0067)*53.491
+
+            struvite.imass['Struvite'] = mixture.imass['P']*\
+                                         self.P_pre_recovery_ratio/\
+                                         30.973762*245.41
+            supply_MgCl2.phase = supply_NH4Cl.phase = base.phase = 's'
+            
+            effluent.copy_like(mixture)
+            effluent.imass['P'] -= struvite.imass['Struvite']*30.973762/245.41
+            effluent.imass['N'] += supply_NH4Cl.imass['NH4Cl']*14.0067/53.491 -\
+                                   struvite.imass['Struvite']*14.0067/245.41
+            effluent.imass['H2O'] = self.F_mass_in - struvite.F_mass -\
+                                    effluent.imass['C'] - effluent.imass['N'] -\
+                                    effluent.imass['P']
+            struvite.phase = 's'    
+                
+            struvite.T = mixture.T
+            effluent.T = mixture.T
+        
+    @property
+    def struvite_P(self):
+        return self.outs[0].imass['Struvite']*30.973762/245.41
+
+    @property
+    def struvite_N(self):
+        return self.struvite_P*14.0067/30.973762
+
+    def _design(self):
+        self.N = ceil(self.HTLmixer.ins[0]._source.WWTP.ins[0].F_vol*2/788.627455/self.V)
+        # 2/788.627455 m3 reactor/m3 wastewater/h (50 MGD ~ 20 m3)
+        self.P = self.ins[0].P
+        Reactor._design(self)
